requirements.txt
asgiref==3.8.1
Django==5.0.4
packaging==24.0
pillow==10.3.0
pytesseract==0.3.10
sqlparse==0.5.0
tesseract==0.1.3
tzdata==2024.1
